Layers for category created, updated

+++ BookRentalApp.Business/Dto/Book/CreateBookDto.cs
@@ -1,10 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
 namespace BookRentalApp.Business.Dto.Book
 {
     public class CreateBookDto
@@ -19,7 +15,7 @@ namespace BookRentalApp.Business.Dto.Book
 
         public int Page { get; set; }
 
-        public int Price { get; set; }
+        public double Price { get; set; }
 
         public int CategoryId { get; set; }
     }
